libbb: implement bb_execXX function family to handle applet executions
This commit implements proxy functions for the exec family of
functions.

bb_execv, bb_execve and bb_execvp were implemented as
proxy functions to their libc counterparts.

applet_execve was implemented as an internal utility
function and is used by bb_execve and bb_execvp to handle applet
invocations.

When called, applet_execve looks for an applet with a matching
name (assuming FEATURE_PREFER_APPLETS is enabled).

When an applet is found, it is executed in one of two ways:
1. If the applet is marked NOEXEC in applets.h, it is
executed using run_noexec_applet_and_exit(). This requires manual
bootstrapping, to mimic the behaviour of the actual exec syscall.
2. If the applet is not marked NOEXEC in applets.h, it is executed
by re-executing the busybox executable with new argv and envp.

When FEATURE_FORCE_APPLETS is enabled, the bb_execXX functions
will fail if an applet is not found, and only applets that are
compiled into the program will be able to run.

When FEATURE_TRY_BASENAME_APPLETS is enabled, applet_execve will
try to find the applet by it's basename (essentially faking a PATH
lookup). This feature, in combination with FEATURE_FORCE_APPLETS,
can be used to force applet execution while mostly maintaining
full functionality in applets that try to execute shells.

These changes allow for better control over external programs
executed by busybox, and provide a unified interface that can
be used in the future to implement further execution limitations.

Moreover, having applet execution logic in one function reduces
code duplication in several applets that attempt to implement
this logic.

Notes:
1. applet_execve copies argv to a new heap-allocated array because
applets are able to modify argv strings. This avoids potential
memory corruptions.
2. applet_execve manually closes FDs marked with FD_CLOEXEC and
resets custom signal handlers to mimic the behaviour of the exec
syscall.
3. BB_EXECVP_or_die was replaced with the bb_execvp_or_die macro.
4. NOMMU targets do not support NOEXEC due to vfork limitations.

Signed-off-by: Nadav Tasher <tashernadav@gmail.com>