libbb: implement bb_system using spawn_and_wait
Implemented bb_system using spawn_and_wait in conjuction with "sh",
to allow bb_system to execute the internal shell when using the
FEATURE_PREFER_APPLETS config option.

This feature requires "sh" to be masked to some shell applet, as
many applets use system to execute commands, and this call should
be handled by the internal shell to allow for further applet
execution.
This requirement is enforced using a new dependency for
FEATURE_PREFER_APPLETS, which is !SH_IS_NONE.

When FEATURE_PREFER_APPLETS is disabled, libc "system()" is used.

Signed-off-by: Nadav Tasher <tashernadav@gmail.com>