httpd: replace execv call with bb_execv and disable FEATURE_HTTPD_CGI when FEATURE_PREFER_APPLETS is enabled
This change replaces the execv in the CGI execution code to bb_execv.
A configuration requirement is added to enforce that applets will
NOT be executed by bb_execv, as this could cause regular requests to
execute unwanted applets instead of regular CGI binaries.

This bb_execv call only serves as a proxy to execv.

Signed-off-by: Nadav Tasher <tashernadav@gmail.com>